requirements.txt
openai-whisper>=20231117
faster-whisper>=1.0.0
torch>=2.0.0
numpy>=1.24.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6



